move 'get-contextual-toolbar'action to the common folder
